ai-seo-writer/requirements.txt
google-generativeai
duckduckgo-search
python-dotenv
beautifulsoup4
requests
rich
